delivery/requirements.txt
google-cloud-bigquery
google-cloud-storage
schema